refactor: share message registry
Move message registry to shared package, so that it can be
shared between the CLI and the toolkit.
