+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/io/index.ts
@@ -1,3 +1,5 @@
 export * from './io-host';
 export * from './io-message';
 export * from './toolkit-action';
+export * from './payloads';
+export * from './messages';
