+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/io/payloads/list.ts
@@ -1,0 +1,5 @@
+import type { StackDetails } from './stack-details';
+
+export interface StackDetailsPayload {
+  stacks: StackDetails[];
+}
