+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/io/payloads/watch.ts
@@ -1,0 +1,29 @@
+
+/**
+ * The computed file watch settings
+ */
+export interface WatchSettings {
+  /**
+   * The directory observed for file changes
+   */
+  readonly watchDir: string;
+  /**
+   * List of include patterns for watching files
+   */
+  readonly includes: string[];
+  /**
+   * List of excludes patterns for watching files
+   */
+  readonly excludes: string[];
+}
+
+export interface FileWatchEvent {
+  /**
+   * The change to the path
+   */
+  readonly event: string;
+  /**
+   * The path that has an observed event
+   */
+  readonly path?: string;
+}
