+++ packages/@aws-cdk/toolkit-lib/lib/actions/bootstrap/index.ts
@@ -227,20 +227,3 @@ export class BootstrapSource {
     };
   }
 }
-
-export interface BootstrapEnvironmentProgress {
-  /**
-   * The total number of environments being deployed
-   */
-  readonly total: number;
-  /**
-   * The count of the environment currently bootstrapped
-   *
-   * This is counting value, not an identifier.
-   */
-  readonly current: number;
-  /**
-   * The environment that's currently being bootstrapped
-   */
-  readonly environment: cxapi.Environment;
-}
