+++ packages/@aws-cdk/toolkit-lib/lib/actions/destroy/index.ts
@@ -1,4 +1,3 @@
-import type { CloudFormationStackArtifact } from '@aws-cdk/cx-api';
 import type { StackSelector } from '../../api/cloud-assembly';
 
 export interface DestroyOptions {
@@ -18,21 +17,4 @@ export interface DestroyOptions {
    * @deprecated has no effect, please implement in IoHost instead
    */
   readonly ci?: boolean;
-}
-
-export interface StackDestroyProgress {
-  /**
-   * The total number of stacks being destroyed
-   */
-  readonly total: number;
-  /**
-   * The count of the stack currently attempted to be destroyed
-   *
-   * This is counting value, not an identifier.
-   */
-  readonly current: number;
-  /**
-   * The stack that's currently being destroyed
-   */
-  readonly stack: CloudFormationStackArtifact;
 }
