+++ packages/@aws-cdk/toolkit-lib/lib/actions/diff/index.ts
@@ -115,23 +115,3 @@ export interface DiffOptions {
    */
   readonly securityOnly?: boolean;
 }
-
-/**
- * Different types of permission related changes in a diff
- */
-export enum PermissionChangeType {
-  /**
-   * No permission changes
-   */
-  NONE = 'none',
-
-  /**
-   * Permissions are broadening
-   */
-  BROADENING = 'broadening',
-
-  /**
-   * Permissions are changed but not broadening
-   */
-  NON_BROADENING = 'non-broadening',
-}
