+++ packages/@aws-cdk/toolkit-lib/lib/actions/index.ts
@@ -1,7 +1,6 @@
 export * from './bootstrap';
 export * from './deploy';
 export * from './destroy';
-export { PermissionChangeType } from './diff';
 export * from './list';
 export * from './rollback';
 export * from './synth';
