+++ packages/@aws-cdk/toolkit-lib/lib/actions/list/index.ts
@@ -1,4 +1,3 @@
-import type { StackDetails } from '../../api/aws-cdk';
 import type { StackSelector } from '../../api/cloud-assembly';
 
 export interface ListOptions {
@@ -6,8 +5,4 @@ export interface ListOptions {
    * Select the stacks
    */
   readonly stacks?: StackSelector;
-}
-
-export interface StackDetailsPayload {
-  stacks: StackDetails[];
 }
