+++ packages/@aws-cdk/toolkit-lib/lib/actions/rollback/index.ts
@@ -1,4 +1,3 @@
-import type { CloudFormationStackArtifact } from '@aws-cdk/cx-api';
 import type { StackSelector } from '../../api/cloud-assembly';
 
 export interface RollbackOptions {
@@ -42,21 +41,4 @@ export interface RollbackOptions {
    * @default true
    */
   readonly validateBootstrapStackVersion?: boolean;
-}
-
-export interface StackRollbackProgress {
-  /**
-   * The total number of stacks being rolled back
-   */
-  readonly total: number;
-  /**
-   * The count of the stack currently attempted to be rolled back
-   *
-   * This is counting value, not an identifier.
-   */
-  readonly current: number;
-  /**
-   * The stack that's currently being rolled back
-   */
-  readonly stack: CloudFormationStackArtifact;
 }
