+++ packages/@aws-cdk/toolkit-lib/lib/api/cloud-assembly/index.ts
@@ -1,4 +1,4 @@
-export * from './context';
+export * from '../../api/shared-public';
 export * from './source-builder';
 export * from './stack-selector';
 export * from './types';
