+++ packages/@aws-cdk/toolkit-lib/lib/api/io/private/index.ts
@@ -1,4 +1,4 @@
-export * from './messages';
+export * from '../../shared-public';
 export * from './io-host-wrappers';
 export * from './timer';
 export * from './sdk-logger';
