+++ packages/@aws-cdk/toolkit-lib/lib/api/io/private/sdk-logger.ts
@@ -1,9 +1,9 @@
 
 import { inspect } from 'util';
 import type { Logger } from '@smithy/types';
-import { IO } from './messages';
 import { replacerBufferWithInfo } from '../../../private/util';
 import type { IoHelper } from '../../shared-private';
+import { IO } from '../../shared-public';
 
 export function asSdkLogger(ioHost: IoHelper): Logger {
   return new class implements Logger {
