+++ packages/@aws-cdk/toolkit-lib/lib/toolkit/index.ts
@@ -1,2 +1,2 @@
 export * from './toolkit';
-export * from './types';
+export * from '../api/shared-public';
