+++ packages/aws-cdk/lib/api/deployments/asset-publishing.ts
@@ -199,6 +199,7 @@ export abstract class BasePublishProgressListener implements IPublishProgressLis
         formatMessage({
           level,
           message: this.getMessage(type, event),
+          data: undefined,
         }),
       );
     }
