+++ packages/aws-cdk/lib/api/stack-events/stack-progress-monitor.ts
@@ -1,24 +1,8 @@
 
 import * as util from 'util';
+import type { StackProgress } from '@aws-cdk/tmp-toolkit-helpers';
 import { StackEvent } from '@aws-sdk/client-cloudformation';
 import { padLeft } from '../../util';
-
-export interface StackProgress {
-  /**
-   * The total number of progress monitored resources.
-   */
-  readonly total?: number;
-
-  /**
-   * The number of completed resources.
-   */
-  readonly completed: number;
-
-  /**
-   * The current progress as a [34/42] string, or just [34] if the total is unknown.
-   */
-  readonly formatted: string;
-}
 
 /**
  * Monitors stack progress.s
