+++ packages/aws-cdk/lib/cli/activity-printer/history.ts
@@ -1,7 +1,7 @@
 import * as util from 'util';
+import type { StackActivity } from '@aws-cdk/tmp-toolkit-helpers';
 import * as chalk from 'chalk';
 import { ActivityPrinterBase, ActivityPrinterProps } from './base';
-import type { StackActivity } from '../../api/stack-events';
 import { padRight } from '../../util';
 
 /**
